refactor(dart): refactor dart structure
refactor dart structure
