--- lib/interfaces/logto_interfaces.dart
@@ -1,2 +1,0 @@
-export 'logto_config.dart';
-export 'oidc_provider_config.dart';
