+++ lib/src/interfaces/logto_interfaces.dart
@@ -1,0 +1,1 @@
+export 'logto_config.dart';
