+++ test/logto_client_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:logto_dart_sdk/interfaces/logto_config.dart';
+import 'package:logto_dart_sdk/src/interfaces/logto_config.dart';
 import 'package:logto_dart_sdk/logto_client.dart';
 
 void main() {
@@ -8,7 +8,8 @@ void main() {
     const String appId = 'foo';
     const String endpoint = 'foo@siverhand.io';
 
-    final config = LogtoConfig(appId: 'foo', endpoint: 'foo@siverhand.io');
+    const config = LogtoConfig(appId: 'foo', endpoint: 'foo@siverhand.io');
+
     final logto = Logto(config);
 
     expect(logto.config.appId, appId);
